requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
edge-tts==6.1.9
httpx==0.26.0
python-dotenv==1.0.1
websockets==12.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community==0.0.21
langchain-openai==0.0.5
langgraph==0.0.26
pydantic==2.6.1
websocket-client==1.7.0
numpy==1.26.4
scipy==1.12.0
aiohttp==3.9.3
soundfile==0.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 测试依赖
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0 